+++ src/TrackingCodeWidget.php
@@ -4,8 +4,6 @@ namespace luya\matomo;
 
 use luya\base\Widget;
 use luya\web\View;
-use luya\matomo\Module;
-use yii\base\InvalidConfigException;
 
 /**
  * Integrate Matomo TrackingCode.
@@ -17,78 +15,56 @@ use yii\base\InvalidConfigException;
  */
 class TrackingCodeWidget extends Widget
 {
-    /**
-     * @var integer The given siteId from the piwik interface.
-     */
-    public $siteId;
+    use ServerAndSiteConfigurationTrait;
     
     /**
-     * @var string The path to the piwik server with schema like `https://piwik.mydomain.com`
+     * @var string Disable cookies for the visitor: https://matomo.org/faq/general/faq_157/
+     * @since 1.0.1
      */
-    public $serverUrl;
+    public $enableCookies = true;
     
     /**
-     * @inheritdoc
+     * @var string Overrides whether the script should be rendered and assigned, as its default disabled when not in production.
+     * @since 1.0.1
      */
-    public function init()
+    public $forceRender = false;
+    
+    public function getJsCode()
     {
-        parent::init();
+        $push = [
+            "_paq.push(['trackPageView']);",
+            "_paq.push(['enableLinkTracking']);"
+        ];
+        
+        if ($this->enableCookies === false) {
+            $push[] = "_paq.push(['disableCookies']);";
+        }
+        
+        return "var _paq = _paq || [];
+".implode(PHP_EOL, $push)."
+(function() {
+    var u=\"".$this->getEnsuredServerUrl()."/\";
+    _paq.push(['setTrackerUrl', u+'piwik.php']);
+    _paq.push(['setSiteId', '".$this->getSiteId()."']);
+    var d=document, g=d.createElement('script'), s=d.getElementsByTagName('script')[0];
+    g.type='text/javascript'; g.async=true; g.defer=true; g.src=u+'piwik.js'; s.parentNode.insertBefore(g,s);
+})();";
+        
+    }
 
-        if ($this->siteId === null) {
-            $this->siteId = $this->module->siteId;
-        }
-        
-        if ($this->serverUrl === null) {
-            $this->serverUrl = $this->module->serverUrl;
-        }
-    }
-    
-    private $_module;
-    
-    /**
-     * Get the matomo module.
-     *
-     * Returns the admin module object instance.
-     *
-     * @return \luya\matomo\admin\Module
-     */
-    public function getModule()
+    public function isTrackingRender()
     {
-        if ($this->_module === null) {
-            $this->_module = Module::getInstance();
-            
-            if ($this->_module === null) {
-                throw new InvalidConfigException("You have either provide the sideId and serverUrl as widget config or configure the luya\matomo\Module in your config.");
-            }
-        }
-        
-        return $this->_module;
-    }
-    
-    public function getEnsuredHostName()
-    {
-        return Module::ensureServerUrl($this->serverUrl);
+        return $this->forceRender || YII_ENV_PROD;
     }
     
     public function run()
     {
         // if not in production, the widget does not render and register anything.
-        if (YII_ENV_PROD) {
-            $this->view->registerJs("
-  var _paq = _paq || [];
-  _paq.push(['trackPageView']);
-  _paq.push(['enableLinkTracking']);
-  (function() {
-    var u=\"".$this->getEnsuredHostName()."/\";
-    _paq.push(['setTrackerUrl', u+'piwik.php']);
-    _paq.push(['setSiteId', '".$this->siteId."']);
-    var d=document, g=d.createElement('script'), s=d.getElementsByTagName('script')[0];
-    g.type='text/javascript'; g.async=true; g.defer=true; g.src=u+'piwik.js'; s.parentNode.insertBefore(g,s);
-  })();
-	", View::POS_HEAD);
+        if ($this->isTrackingRender()) {
+            $this->view->registerJs($this->getJsCode(), View::POS_HEAD);
         
             return $this->render('trackingcode', [
-                'hostName' => $this->getEnsuredHostName(),
+                'hostName' => $this->getEnsuredServerUrl(),
                 'siteId' => $this->siteId,
             ]);
         }