+++ CHANGELOG.md
@@ -2,6 +2,11 @@
 
 All notable changes to this project will be documented in this file. This project make usage of the [Yii Versioning Strategy](https://github.com/yiisoft/yii2/blob/master/docs/internals/versions.md). In order to read more about upgrading and BC breaks have a look at the [UPGRADE Document](UPGRADE.md).
 
+## 1.0.1 (in progress)
+
++ [#4](https://github.com/luyadev/luya-module-matomo/issues/4) Widget for opt out iframe.
++ [#5](https://github.com/luyadev/luya-module-matomo/issues/5) Option to disable tracking with cookies.
+
 ## 1.0.0 (5. March 2018)
 
 + First stable release